initialModuleState()
